separando los componentes en archivos individuales
