+++ redux-counter/hook-counter/src/App.jsx
@@ -1,31 +1,17 @@
-/* eslint-disable react/prop-types */
 import './App.css'
-
-import CounterContext from './CounterContext'
-
-import { useContext } from 'react'
-
-const Display = () => {
-  const [counter] = useContext(CounterContext)
-  return <div>{counter}</div>
-}
-
-const Button = ({ type, label }) => {
-  // eslint-disable-next-line no-unused-vars
-  const [counter, dispatch] = useContext(CounterContext)
-  return <button onClick={() => dispatch({ type })}>{label}</button>
-}
+import Display from './components/Display'
+import Button from './components/Button'
 
 const App = () => {
   return (
-    <>
+    <div>
       <Display />
       <div>
         <Button type="INC" label="+" />
         <Button type="DEC" label="-" />
         <Button type="ZERO" label="0" />
       </div>
-    </>
+    </div>
   )
 }
