+++ redux-counter/hook-counter/src/components/Display.jsx
@@ -1,0 +1,9 @@
+import { useContext } from 'react'
+import CounterContext from '../CounterContext'
+
+const Display = () => {
+  const [counter] = useContext(CounterContext)
+  return <div>{counter}</div>
+}
+
+export default Display
